Add files via upload
 Vehicular Telematics has become one of the 
emerging research topics for intelligent transport systems. 
In this paper we present a novel approach to contour and 
extract characters from License plates with 90% efficiency. 
The aim of this proposed algorithm is to find the connected 
components in a vehicle License plate to extract characters 
as the isolated characters belong to one set of connected 
components. We have also compared the results with other 
existing algorithms like Moore neighborhood pixel, and 
modified Moore neighborhood pixel etc. and found our 
proposed algorithm to be 20-27% more efficient than the 
existing algorithms.